+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.2.0] - 2025-10-10
+
+* add the ability to create a hidden access point
+* remove WiFiManager.debugMemory() function as non-essential
+
 ## [1.1.0] - 2025-03-06
 
 * add user callback function to first successful connection event